requirements.txt
tweepy-4.4.0
schedule-1.1.0